refactor(helper): update helper subdir structure
