+++ src/components/ShortKeys.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { IoReturnDownBackSharp } from 'react-icons/io5';
 import { Shortcut } from '../enums/Shortcut.enum';
-import { cn } from '../helpers/cn.helper';
+import { cn } from '../helpers/className/cn.helper';
 
 const ShortKeys: React.FC<{ keys: Shortcut[]; className?: string }> = ({
   keys,
