+++ src/helpers/className/isDomainValid.helper.ts
